Rename private modules in optuna.pruners.
