+++ optuna/pruners/__init__.py
@@ -1,11 +1,11 @@
 import optuna
-from optuna.pruners.base import BasePruner  # NOQA
-from optuna.pruners.hyperband import HyperbandPruner  # NOQA
-from optuna.pruners.median import MedianPruner  # NOQA
-from optuna.pruners.nop import NopPruner  # NOQA
-from optuna.pruners.percentile import PercentilePruner  # NOQA
-from optuna.pruners.successive_halving import SuccessiveHalvingPruner  # NOQA
-from optuna.pruners.threshold import ThresholdPruner  # NOQA
+from optuna.pruners._base import BasePruner  # NOQA
+from optuna.pruners._hyperband import HyperbandPruner  # NOQA
+from optuna.pruners._median import MedianPruner  # NOQA
+from optuna.pruners._nop import NopPruner  # NOQA
+from optuna.pruners._percentile import PercentilePruner  # NOQA
+from optuna.pruners._successive_halving import SuccessiveHalvingPruner  # NOQA
+from optuna.pruners._threshold import ThresholdPruner  # NOQA
 
 
 def _filter_study(
