+++ optuna/pruners/_nop.py
