+++ tests/pruners_tests/test_percentile.py
@@ -3,7 +3,7 @@ import math
 import pytest
 
 import optuna
-from optuna.pruners import percentile
+from optuna.pruners import _percentile
 from optuna.study import StudyDirection
 from optuna.trial import TrialState
 from optuna import type_checking
@@ -140,7 +140,7 @@ def test_get_best_intermediate_result_over_steps(direction_expected):
     trial_empty = study._storage.get_trial(trial_id_empty)
 
     with pytest.raises(ValueError):
-        percentile._get_best_intermediate_result_over_steps(trial_empty, direction)
+        _percentile._get_best_intermediate_result_over_steps(trial_empty, direction)
 
     # Input value has no NaNs but float values.
     trial_id_float = study._storage.create_new_trial(study._study_id)
@@ -148,7 +148,7 @@ def test_get_best_intermediate_result_over_steps(direction_expected):
     trial_float.report(0.1, step=0)
     trial_float.report(0.2, step=1)
     frozen_trial_float = study._storage.get_trial(trial_id_float)
-    assert expected == percentile._get_best_intermediate_result_over_steps(
+    assert expected == _percentile._get_best_intermediate_result_over_steps(
         frozen_trial_float, direction
     )
 
@@ -158,7 +158,7 @@ def test_get_best_intermediate_result_over_steps(direction_expected):
     trial_float_nan.report(0.3, step=0)
     trial_float_nan.report(float("nan"), step=1)
     frozen_trial_float_nan = study._storage.get_trial(trial_id_float_nan)
-    assert 0.3 == percentile._get_best_intermediate_result_over_steps(
+    assert 0.3 == _percentile._get_best_intermediate_result_over_steps(
         frozen_trial_float_nan, direction
     )
 
@@ -168,7 +168,7 @@ def test_get_best_intermediate_result_over_steps(direction_expected):
     trial_nan.report(float("nan"), step=0)
     frozen_trial_nan = study._storage.get_trial(trial_id_nan)
     assert math.isnan(
-        percentile._get_best_intermediate_result_over_steps(frozen_trial_nan, direction)
+        _percentile._get_best_intermediate_result_over_steps(frozen_trial_nan, direction)
     )
 
 
@@ -186,7 +186,7 @@ def test_get_percentile_intermediate_result_over_trials():
             with pytest.raises(ValueError):
                 _all_trials = _study._storage.get_all_trials(_study._study_id)
                 _direction = _study._storage.get_study_direction(_study._study_id)
-                percentile._get_percentile_intermediate_result_over_trials(
+                _percentile._get_percentile_intermediate_result_over_trials(
                     _all_trials, _direction, step, 25
                 )
 
@@ -206,7 +206,7 @@ def test_get_percentile_intermediate_result_over_trials():
     study = setup_study(9, intermediate_values)
     all_trials = study._storage.get_all_trials(study._study_id)
     direction = study._storage.get_study_direction(study._study_id)
-    assert 0.3 == percentile._get_percentile_intermediate_result_over_trials(
+    assert 0.3 == _percentile._get_percentile_intermediate_result_over_trials(
         all_trials, direction, 0, 25.0
     )
 
@@ -217,7 +217,7 @@ def test_get_percentile_intermediate_result_over_trials():
     study = setup_study(9, intermediate_values)
     all_trials = study._storage.get_all_trials(study._study_id)
     direction = study._storage.get_study_direction(study._study_id)
-    assert 0.2 == percentile._get_percentile_intermediate_result_over_trials(
+    assert 0.2 == _percentile._get_percentile_intermediate_result_over_trials(
         all_trials, direction, 1, 25.0
     )
 
@@ -239,5 +239,5 @@ def test_get_percentile_intermediate_result_over_trials():
     all_trials = study._storage.get_all_trials(study._study_id)
     direction = study._storage.get_study_direction(study._study_id)
     assert math.isnan(
-        percentile._get_percentile_intermediate_result_over_trials(all_trials, direction, 2, 75)
+        _percentile._get_percentile_intermediate_result_over_trials(all_trials, direction, 2, 75)
     )
